backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.1
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
anthropic==0.42.0
openai==1.58.1
chromadb==0.5.23
tiktoken==0.8.0
apscheduler==3.10.4
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.1
python-multipart==0.0.20
httpx==0.28.1
slowapi==0.1.9
email-validator==2.3.0

